r.pack/r.unpack: fix packing/unpacking vrt map if some of tiled rasters doesn't have color table (#1116)
* r.pack: fix packing vrt map if some of tiled rasters doesn't have color table
* r.unpack: fix unpacking vrt map with tiled raster maps

+++ scripts/r.pack/r.pack.py
@@ -123,15 +123,16 @@ def main():
                 if not os.path.exists(f_tmp_dir):
                     os.mkdir(f_tmp_dir)
                 path = os.path.join(vrt_files[f], element, f)
-                grass.debug("copying vrt file {}".format(path))
-                if os.path.isfile(path):
-                    shutil.copyfile(
-                        path, os.path.join(f_tmp_dir, element),
-                    )
-                else:
-                    shutil.copytree(
-                        path, os.path.join(f_tmp_dir, element),
-                    )
+                if os.path.exists(path):
+                    grass.debug("copying vrt file {}".format(path))
+                    if os.path.isfile(path):
+                        shutil.copyfile(
+                            path, os.path.join(f_tmp_dir, element),
+                        )
+                    else:
+                        shutil.copytree(
+                            path, os.path.join(f_tmp_dir, element),
+                        )
 
     if not os.listdir(tmp_dir):
         grass.fatal(_("No raster map components found"))
